+++ src/main/java/chess/service/ChessGameService.java
@@ -1,6 +1,7 @@
 package chess.service;
 
 import chess.dao.ChessGameDao;
+import chess.dao.PieceDao;
 import chess.domain.ChessGame;
 import chess.domain.GameResult;
 import chess.domain.generator.BlackGenerator;
@@ -10,6 +11,7 @@ import chess.domain.player.Team;
 import chess.domain.position.Position;
 import chess.dto.ChessGameDto;
 import chess.dto.ChessGameUpdateDto;
+import chess.dto.PieceDto;
 import chess.dto.StatusDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
@@ -18,9 +20,11 @@ import org.springframework.stereotype.Service;
 public class ChessGameService {
 
     private final ChessGameDao chessGameDao;
+    private final PieceDao pieceDao;
 
-    public ChessGameService(ChessGameDao chessGameDao) {
+    public ChessGameService(ChessGameDao chessGameDao, PieceDao pieceDao) {
         this.chessGameDao = chessGameDao;
+        this.pieceDao = pieceDao;
     }
 
     public ChessGameDto createNewChessGame(final String gameName) {
@@ -31,8 +35,15 @@ public class ChessGameService {
         final Player whitePlayer = new Player(new WhiteGenerator(), Team.WHITE);
         final Player blackPlayer = new Player(new BlackGenerator(), Team.BLACK);
         final ChessGame chessGame = new ChessGame(whitePlayer, blackPlayer);
-        chessGameDao.createNewChessGame(chessGame, gameName);
+        saveNewChessGame(chessGame, gameName);
         return ChessGameDto.of(chessGame, gameName);
+    }
+
+    private void saveNewChessGame(final ChessGame chessGame, final String gameName) {
+        chessGameDao.saveChessGame(gameName, chessGame.getTurn());
+        final int chessGameId = chessGameDao.findChessGameIdByName(gameName);
+        pieceDao.savePieces(chessGame.getCurrentPlayer(), chessGameId);
+        pieceDao.savePieces(chessGame.getOpponentPlayer(), chessGameId);
     }
 
     public StatusDto findStatus(final String gameName) {
@@ -45,8 +56,8 @@ public class ChessGameService {
 
     public StatusDto finishGame(final String gameName) {
         final StatusDto status = findStatus(gameName);
-        final int gameId = findGameIdByGameName(gameName);
-        chessGameDao.deletePieces(gameId);
+        final int gameId = chessGameDao.findChessGameIdByName(gameName);
+        pieceDao.deletePieces(gameId);
         chessGameDao.deleteChessGame(gameId);
         return status;
     }
@@ -57,24 +68,33 @@ public class ChessGameService {
     }
 
     public ChessGameDto move(final String gameName, final String current, final String destination) {
-        final int gameId = findGameIdByGameName(gameName);
+        final int gameId = chessGameDao.findChessGameIdByName(gameName);
         final ChessGame chessGame = findGameByName(gameName);
         final Player currentPlayer = chessGame.getCurrentPlayer();
         final Player opponentPlayer = chessGame.getOpponentPlayer();
         chessGame.move(currentPlayer, opponentPlayer, new Position(current), new Position(destination));
         chessGameDao.updateGameTurn(gameId, chessGame.getTurn());
-        chessGameDao.updatePiece(gameId, current, destination, currentPlayer.getTeamName(),
+        updatePiece(gameId, current, destination, currentPlayer.getTeamName(),
                 opponentPlayer.getTeamName());
         return ChessGameDto.of(chessGame, gameName);
     }
 
+    public void updatePiece(final int gameId, final String current, final String destination,
+                            final String currentTeam, final String opponentTeam) {
+        pieceDao.deletePieceByGameIdAndPositionAndTeam(gameId, destination, opponentTeam);
+        pieceDao.updatePiecePositionByGameId(gameId, current, destination, currentTeam);
+    }
+
     private ChessGame findGameByName(final String gameName) {
-        final int chessGameId = findGameIdByGameName(gameName);
-        final ChessGameUpdateDto gameUpdateDto = chessGameDao.findChessGame(chessGameId);
+        final int chessGameId = chessGameDao.findChessGameIdByName(gameName);
+        final ChessGameUpdateDto gameUpdateDto = findChessGame(chessGameId);
         return ChessGame.of(gameUpdateDto);
     }
 
-    private int findGameIdByGameName(final String gameName) {
-        return chessGameDao.findChessGameIdByName(gameName);
+    public ChessGameUpdateDto findChessGame(final int chessGameId) {
+        final String turn = chessGameDao.findCurrentTurn(chessGameId);
+        final List<PieceDto> whitePieces = pieceDao.findAllPieceByIdAndTeam(chessGameId, Team.WHITE.getName());
+        final List<PieceDto> blackPieces = pieceDao.findAllPieceByIdAndTeam(chessGameId, Team.BLACK.getName());
+        return new ChessGameUpdateDto(turn, whitePieces, blackPieces);
     }
 }
